+++ test/lib/configManager/.gitignore
@@ -1,0 +1,1 @@
+init.yaml
